Support multiple m4b source files

+++ audiobook_tools/m4b2mp3.py
@@ -7,11 +7,18 @@ from helpers import *
 
 def main():
     parser = common_parser("m4b2mp3")
-    parser.add_argument("m4b", type=str, help="m4b file")
+    parser.add_argument("m4b", type=str, help="m4b file", nargs="+")
     args = parser.parse_args()
     logger.setLevel(logging.getLevelName(args.log_level.upper()))
-    input_file = abspath(args.m4b)
-    tags = get_m4b_tags(input_file)
+    input_files = []
+    for file_pattern in args.m4b:
+        if "*" in file_pattern:
+            input_files.extend(expand_glob(file_pattern))
+        else:
+            input_files.append(abspath(file_pattern))
+    log(f"There are {len(input_files)} m4b files to process: {input_files}", debug)
+    # input_file = abspath(args.m4b)
+    tags = get_m4b_tags(input_files[0])
     output_dir = abspath(os.path.join(os.getcwd(), f"{tags['author']}/{tags['title']}"))
     output_dir = abspath(get_input("Output directory", f"{output_dir}"))
     output_image = os.path.join(output_dir, "cover.jpg")
@@ -20,23 +27,37 @@ def main():
     output_metadata = abspath(get_input("Output metadata file", f"{output_metadata}"))
     os.makedirs(output_dir, exist_ok=True)
     run(
-        f"ffmpeg -y -i {shlex.quote(input_file)} -an -c:v copy {shlex.quote(output_image)}"
+        f"ffmpeg -y -i {shlex.quote(input_files[0])} -an -c:v copy {shlex.quote(output_image)}"
     )
-    chapters = json.loads(
-        run(
-            f"ffprobe -v quiet -print_format json -show_chapters {shlex.quote(input_file)}"
-        )
-    ).get("chapters", {})
-    print("Splitting m4b file into chapter tracks...")
-    for chapter in chapters:
-        chapter["track"] = str(int(chapter["id"]) + 1).zfill(len(str(len(chapters))))
-        chapter["section"] = "1"
-        chapter["chapter"] = chapter["tags"]["title"]
-        chapter.pop("tags")
-        m4b_output_file = os.path.join(output_dir, f"{chapter['id']}.m4b")
-        run(
-            f"ffmpeg -y -i '{input_file}' -sn -vn -codec copy -ss {chapter['start_time']} -to {chapter['end_time']} '{m4b_output_file}'"
-        )
+    if len(input_files) == 1:
+        chapters = json.loads(
+            run(
+                f"ffprobe -v quiet -print_format json -show_chapters {shlex.quote(input_files[0])}"
+            )
+        ).get("chapters", {})
+        print("Splitting m4b file into chapter tracks...")
+        for chapter in chapters:
+            chapter["track"] = str(int(chapter["id"]) + 1).zfill(
+                len(str(len(chapters)))
+            )
+            chapter["section"] = "1"
+            chapter["chapter"] = chapter["tags"]["title"]
+            chapter.pop("tags")
+            m4b_output_file = os.path.join(output_dir, f"{chapter['id']}.m4b")
+            run(
+                f"ffmpeg -y -i '{input_files[0]}' -sn -vn -codec copy -ss {chapter['start_time']} -to {chapter['end_time']} '{m4b_output_file}'"
+            )
+    else:
+        chapters = []
+        for idx, input_file in enumerate(input_files):
+            log(f"Processing file: {input_file}", debug)
+            chapter_tags = get_m4b_tags(input_file, prompt=False)
+            chapter = {}
+            chapter["file"] = input_file
+            chapter["track"] = str(idx + 1).zfill(len(str(len(input_files))))
+            chapter["section"] = "1"
+            chapter["chapter"] = chapter_tags["title"]
+            chapters.append(chapter)
     yaml.dump(
         merge_dicts(tags, {"chapters": chapters}),
         open(output_metadata, "w"),
@@ -64,9 +85,16 @@ def main():
         mp3_output_file = os.path.join(
             output_dir, (render(file_name_format, merge_dicts(mp3_metadata, chapter)))
         )
-        run(
-            f"ffmpeg -y -i '{input_file}' -sn -vn -codec:a libmp3lame -qscale:a 2 -ss {chapter['start_time']} -to {chapter['end_time']} {shlex.quote(mp3_output_file)}"
-        )
+        if len(input_files) == 1:
+            input_file = input_files[0]
+            run(
+                f"ffmpeg -y -i '{input_file}' -map_metadata -1 -sn -vn -codec:a libmp3lame -qscale:a 2 -ss {chapter['start_time']} -to {chapter['end_time']} {shlex.quote(mp3_output_file)}"
+            )
+        else:
+            input_file = chapter["file"]
+            run(
+                f"ffmpeg -y -i '{input_file}' -map_metadata -1 -sn -vn -codec:a libmp3lame -qscale:a 2 {shlex.quote(mp3_output_file)}"
+            )
         total_tracks_in_section = len(
             [c for c in mp3_metadata["chapters"] if c["section"] == chapter["section"]]
         )
@@ -87,10 +115,11 @@ def main():
         )
 
     # Remove m4b files
-    print("Removing m4b files...")
-    for chapter in mp3_metadata["chapters"]:
-        m4b_output_file = os.path.join(output_dir, f"{chapter['id']}.m4b")
-        os.remove(m4b_output_file)
+    if len(input_files) == 1:
+        print("Removing m4b files...")
+        for chapter in mp3_metadata["chapters"]:
+            m4b_output_file = os.path.join(output_dir, f"{chapter['id']}.m4b")
+            os.remove(m4b_output_file)
 
 
 if __name__ == "__main__":
